requirements.txt
Flask==2.3.3
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7